+++ examples/svelte-runes/src/lib/store.svelte.ts
@@ -10,7 +10,7 @@ type Counter = {
   nested3: Option<{ foo: { bar: { baz: Option<number> } } }>;
 };
 
-const counter: Counter = {
+const defaultValue: Counter = {
   counter: 0,
   counter2: 0,
   counter3: 0,
@@ -19,7 +19,7 @@ const counter: Counter = {
   nested3: { foo: { bar: { baz: 0 } } },
 };
 
-export const store = new RuneStore('counter-store', counter, {
+export const store = new RuneStore('counter-store', defaultValue, {
   filterKeys: ['counter', 'nested', 'nested2', 'nested3'],
   filterKeysStrategy: 'pick',
   saveOnExit: true,